requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
beautifulsoup4==4.12.2